fix typo, add font etc
fixed the following:
Does not include labels for each form field.
Labels are not properly associated with its form field.
Both radio buttons can be selected at the same time.

I just noticed a couple of small issues:

At mobile which doesn't quite match the mockup, and that is for your zipcode field as its the only one that doesn't stack. The field should be on a separate line below the label like the mockup and the rest of the fields.
It still doesn't look like your using a google font for the project. I noticed you are referencing "Nunito" but you have not got a link to that in the HTML for example. Remember, to use fonts you must either link to the font within the HTML or import it through CSS, or download the font and reference it in your project.